2DAM_DI_TEMA_7_CODIGO/BolT7/src/test/java/vista/PopupMenuDemoTest.java
package vista;

import static org.junit.jupiter.api.Assertions.assertNotNull;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.TestInstance;

//La anotación @TestInstance se utiliza para especificar el ciclo de vida de las instancias de prueba.
@TestInstance(TestInstance.Lifecycle.PER_METHOD)
public class PopupMenuDemoTest {

	
	private PopupMenuDemo popupMenuDemo;

    // Este método se ejecuta antes de cada método de prueba.
    @BeforeEach
    void setUp() {
        popupMenuDemo = new PopupMenuDemo();
    }

    // Este método se ejecuta después de cada método de prueba.
    @AfterEach
    void tearDown() {
        popupMenuDemo = null;
    }

    // Prueba para verificar la creación del menú bar.
    @Test
    void testCreateMenuBar() {
        assertNotNull(popupMenuDemo.createMenuBar());
        // Se pueden agregar más aserciones según sea necesario para verificar la creación del menú.
    }

    // Prueba para verificar la creación del contenido del panel.
    @Test
    void testCreateContentPane() {
        assertNotNull(popupMenuDemo.createContentPane());
        // Se pueden agregar más aserciones según sea necesario para verificar la creación del contenido.
    }

    // Prueba para verificar la creación del menú emergente.
    @Test
    void testCreatePopupMenu() {
        assertNotNull(popupMenuDemo.createPopupMenu());
        // Se pueden agregar más aserciones según sea necesario para verificar la creación del menú emergente.
    }

    // Se pueden agregar más pruebas para probar la interacción de la aplicación con los eventos, menús, etc.

	
}
